Pathfinding
Created function to convert the DebugRestaurant map to an array of booleans, 1 = unwalkable tile else walkable.
Introduction to a Diagonal Heuristic to achieve more human behaviour on the pathfinding. (Old code is commented, I had a decent small heuristic that I can still use).
Node class cleanup. (For some reason had an error because the nodes array in graphgrid was in private, had to change to publci).
GraphGrid class cleanup.